Fix ci & remove dead code
