+++ src/kernel/mod.rs
@@ -1,7 +1,6 @@
 use async_trait::async_trait;
 use bytes::Bytes;
 use fslock::LockFile;
-use itertools::Itertools;
 use serde::{Deserialize, Serialize};
 use std::ffi::OsStr;
 use std::path::Path;
@@ -44,11 +43,6 @@ pub trait Storage: Send + Sync + 'static + Sized {
     /// 通过键删除键值对
     async fn remove(&self, key: &[u8]) -> Result<()>;
 
-    /// 并行批量执行
-    /// TODO need refactor
-    // #[inline]
-    // async fn join(&self, vec_cmd: Vec<CommandData>) -> Result<Vec<Option<Vec<u8>>>> {}
-
     async fn size_of_disk(&self) -> Result<u64>;
 
     async fn len(&self) -> Result<usize>;
@@ -62,55 +56,6 @@ pub enum CommandData {
     Set { key: Vec<u8>, value: Vec<u8> },
     Remove { key: Vec<u8> },
     Get { key: Vec<u8> },
-}
-
-pub(crate) struct ByteUtils;
-
-impl ByteUtils {
-    /// 从u8的slice中前四位获取数据的长度
-    pub(crate) fn from_4_bit_with_start(len_u8: &[u8]) -> usize {
-        usize::from(len_u8[3])
-            | usize::from(len_u8[2]) << 8
-            | usize::from(len_u8[1]) << 16
-            | usize::from(len_u8[0]) << 24
-    }
-
-    /// 返回字节数组Vec与对应的字节数组长度Vec
-    ///
-    /// bytes必须使用'ByteUtils::tag_with_head'进行标记
-    pub(crate) fn sharding_tag_bytes(bytes: &[u8]) -> Vec<&[u8]> {
-        let mut vec_cmd_u8 = Vec::new();
-        let mut last_pos = 0;
-
-        if bytes.len() < 4 {
-            return vec_cmd_u8;
-        }
-
-        loop {
-            let pos = last_pos + 4;
-            if pos >= bytes.len() {
-                break;
-            }
-            let len_u8 = &bytes[last_pos..pos];
-            let len = Self::from_4_bit_with_start(len_u8);
-            if len < 1 || len > bytes.len() {
-                break;
-            }
-
-            last_pos += len + 4;
-            vec_cmd_u8.push(&bytes[pos..last_pos]);
-        }
-
-        vec_cmd_u8
-    }
-
-    /// 标记bytes以支持'ByteUtils::sharding_tag_bytes'方法
-    pub(crate) fn tag_with_head(mut bytes: Vec<u8>) -> Vec<u8> {
-        let i = bytes.len();
-        let mut vec_head = vec![(i >> 24) as u8, (i >> 16) as u8, (i >> 8) as u8, i as u8];
-        vec_head.append(&mut bytes);
-        vec_head
-    }
 }
 
 impl CommandData {
